feat(scaner): add simple char scanner

+++ rox/src/rox/token.rs
@@ -6,11 +6,11 @@ pub struct Token {
     pub token_type: TokenType,
     pub lexeme: String,
     pub literal: String, // NOTE: original is `Object`
-    pub line: usize,
+    pub line: i32,
 }
 
 impl Token {
-    pub fn new(token_type: TokenType, lexeme: String, literal: String, line: usize) -> Token {
+    pub fn new(token_type: TokenType, lexeme: String, literal: String, line: i32) -> Token {
         Token {
             token_type,
             lexeme,
